requirements.txt
aniso8601==8.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
Pillow~=8.4.0
pytz==2020.5
requests==2.25.1
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
simplejson~=3.6.0
WTForms~=3.0.0
opencv-python~=4.5.5.62
numpy~=1.21.5
scipy~=1.7.3